[lld][LoongArch] Define _GLOBAL_OFFSET_TABLE_ when not found in symtab
Fix the issuse similar to https://github.com/llvm/llvm-project/issues/38982.

_GLOBAL_OFFSET_TABLE_ is generated in LoongArch llvm backend when
accessing global variable. It is fine for normal compiliation and
linking, but it does not work for the LTO scenario because lld has an
ordering problem of when the linker defined symbol `_GLOBAL_OFFSET_TABLE_`
is created and when the output ELF file from LTO is loaded.

This patch fixes this issue by defining the `_GLOBAL_OFFSET_TABLE_`
symbol when it is not found in symtab. (Only for LoongArch)